moved command interpretation to interpreter
