+++ src/main/java/net/scottnotfound/clara/lang/Interpreter.java
@@ -1,11 +1,13 @@
 package net.scottnotfound.clara.lang;
 
+import net.scottnotfound.clara.Clara;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Interpreter implements IExprVisitor<Object>, IStmtVisitor<Void> {
+public class Interpreter implements IExprVisitor<Object>, IStmtVisitor<Void>, IArgVisitor<Void>, ICmdVisitor<Void> {
 
     private Environment globals = new Environment();
     private Environment environment = globals;
@@ -311,6 +313,70 @@ public class Interpreter implements IExprVisitor<Object>, IStmtVisitor<Void> {
     public Void visitStmt(Stmt.While stmt) {
         while (isTruthy(evaluateExpression(stmt.condition))) {
             executeStatement(stmt.body);
+        }
+        return null;
+    }
+
+    @Override
+    public Void visitArg(Arg.Argument arg) {
+        return null;
+    }
+
+    @Override
+    public Void visitArg(Arg.Flag arg) {
+        return null;
+    }
+
+    @Override
+    public Void visitArg(Arg.Parameter arg) {
+        return null;
+    }
+
+    @Override
+    public Void visitCmd(Cmd.Default cmd) {
+
+        if (cmd.command == null) {
+
+        } else {
+
+            switch (cmd.command.lexeme) {
+                case ("reaction") : {
+
+                }
+                default: {
+
+                }
+            }
+
+        }
+
+
+        return null;
+    }
+
+    @Override
+    public Void visitCmd(Cmd.Exit cmd) {
+        Clara.shutdown(0);
+        return null;
+    }
+
+    @Override
+    public Void visitCmd(Cmd.Help cmd) {
+
+        if (cmd.command == null) {
+            // only "help" was entered
+            System.out.println("No commands available yet.");
+        } else {
+
+            switch (cmd.command.lexeme) {
+
+
+                default: {
+                    System.out.println("Command not yet implemented.");
+                }
+
+            }
+
         }
         return null;
     }
